pyproject.toml
[tool.poetry]
name = "mit-learn"
version = "0.71.0"
description = "Search index for use with other MIT applications."
license = "BSD-3"
readme = "README.md"
packages = []
authors  = ["MIT ODL"]


[tool.poetry.dependencies]
python = "3.12.6"
attrs = "^24.0.0"
base36 = "^0.1.1"
beautifulsoup4 = "^4.8.2"
boto3 = "^1.26.155"
cairosvg = "^2.7.0"
celery = "^5.3.1"
cffi = "^1.15.1"
cryptography = "^43.0.0"
dj-database-url = "^2.0.0"
dj-static = "^0.0.6"
Django = "4.2.16"
django-anymail = {extras = ["mailgun"], version = "^11.0"}
django-bitfield = "^2.2.0"
django-cache-memoize = "^0.2.0"
django-cors-headers = "^4.0.0"
django-filter = "^2.4.0"
django-guardian = "^2.4.0"
django-hijack = "^3.4.1"
django-imagekit = "^5.0.0"
django-ipware = "^7.0.0"
django-json-widget = "^2.0.0"
django-redis = "^5.2.0"
django-server-status = "^0.7.0"
django-storages = "^1.13.2"
djangorestframework = "^3.14.0"
drf-jwt = "^1.19.2"
drf-spectacular = "^0.27.0"
feedparser = "^6.0.10"
google-api-python-client = "^2.89.0"
html5lib = "^1.1"
ipython = "^8.14.0"
jedi = "^0.19.0"
lxml = "^5.0.0"
markdown2 = "^2.4.8"
nested-lookup = "^0.2.25"
ocw-data-parser = "^0.35.1"
opensearch-dsl = "^2.0.0"
opensearch-py = "^2.0.0"
psycopg2 = "^2.9.6"
pygithub = "^2.0.0"
python-dateutil = "^2.8.2"
python-rapidjson = "^1.8"
pyyaml = "^6.0.0"
redis = "^5.0.0"
requests = "^2.31.0"
sentry-sdk = "^2.13.0"
social-auth-app-django = "^5.2.0"
static3 = "^0.7.0"
tika = "^2.6.0"
tldextract = "^2.2.0"
toolz = "^1.0.0"
ulid-py = "^0.2.0"
urllib3 = "^2.0.0"
uwsgi = "^2.0.21"
wrapt = "^1.14.1"
xbundle = "^0.3.1"
social-auth-core = {extras = ["openidconnect"], version = "^4.4.2"}
nh3 = "^0.2.14"
retry2 = "^0.9.5"
pluggy = "^1.3.0"
named-enum = "^1.4.0"
drf-nested-routers = "^0.94.0"
django-scim2 = "^0.19.1"
django-oauth-toolkit = "^2.3.0"
youtube-transcript-api = "^0.6.2"
posthog = "^3.5.0"
ruff = "0.7.3"
dateparser = "^1.2.0"
uwsgitop = "^0.12"
pytest-lazy-fixtures = "^1.1.1"
pycountry = "^24.6.1"
qdrant-client = {extras = ["fastembed"], version = "^1.12.0"}
openai = "^1.54.4"
llama-index-agent-openai = "^0.3.4"
llama-index-llms-openai = "^0.2.16"
llama-index = "^0.11.23"
channels = {extras = ["daphne"], version = "^4.2.0"}


[tool.poetry.group.dev.dependencies]
bpython = "^0.24"
ddt = "^1.6.0"
django-debug-toolbar = "^4.1.0"
factory_boy = "^3.3.0"
faker = "^30.0.0"
ipdb = "^0.13.13"
moto = "^5.0.0"
nplusone = "^1.0.0"
pdbpp = "^0.10.3"
pytest = "^8.0.0"
pytest-cov = "^5.0.0"
pytest-django = "^4.5.2"
pytest-env = "^1.0.0"
pytest-freezegun = "^0.4.2"
pytest-mock = "^3.10.0"
responses = "^0.25.0"
ruff = "^0.7.0"
safety = "^2.3.4"
semantic-version = "^2.10.0"
freezegun = "^1.4.0"
pytest-xdist = { version = "^3.6.1", extras = ["psutil"] }
anys = "^0.3.0"
locust = "^2.31.2"


[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88
lint.select = [
    "A",  # flake8-builtins
    # "AIR",  # Airflow
    # "ANN",  # flake8-annotations
    "ARG",  # flake8-unused-arguments
    # "ASYNC",  # flake8-async
    "B",  # flake8-bugbear
    "BLE",  # flake8-blind-except
    "C4",  # flake8-comprehensions
    "C90",  # mccabe
    # "COM",  # flake8-commas
    "CPY",  # flake8-copyright
    "D",  # pydocstyle
    "DJ",  # flake8-django
    "DTZ",  # flake8-datetimez
    "E",  # Pycodestyle Error
    "EM",  # flake8-errmsg
    "ERA",  # eradicate
    "EXE",  # flake8-executable
    "F",  # Pyflakes
    "FA",  # flake8-future-annotations
    "FBT",  # flake8-boolean-trap
    "FIX",  # flake8-fixme
    "FLY",  # flynt
    # "FURB",  # refurb
    "G",  # flake8-logging-format
    "I",  # isort
    "ICN",  # flake8-import-conventions
    "INP",  # flake8-no-pep420
    "INT",  # flake8-gettext
    "ISC",  # flake8-implicit-str-concat
    "N",  # pep8-naming
    # "NPY",  # NumPy-specific rules
    # "PD",  # pandas-vet
    "PERF",  # Perflint
    "PGH",  # pygrep-hooks
    "PIE",  # flake8-pie
    "PL",  # Pylint
    "PT",  # flake8-pytest-style
    "PTH",  # flake8-use-pathlib
    "PYI",  # flake8-pyi
    "Q",  # flake8-quotes
    "RET",  # flake8-return
    "RSE",  # flake8-raise
    "RUF",  # Ruff-specific rules
    "S",  # flake8-bandit
    "SIM",  # flake8-simplify
    "SLF",  # flake8-self
    "SLOT",  # flake8-slots
    "T10",  # flake8-debugger
    "T20",  # flake8-print
    "TCH",  # flake8-type-checking
    "TD",  # flake8-todos
    "TID",  # flake8-tidy-imports
    "TRY",  # tryceratops
    "UP",  # pyupgrade
    "W",  # Pycodestyle Warning
    "YTT",  # flake8-2020
]
lint.ignore = [
    "B008",
    "B905",
    "D104",
    "D200",
    "D202",
    "D205",
    "D301",
    "D400",
    "N803",
    "N806",
    "N999",
    "PIE804",
    "RET505",
    "RET506",
    "RET507",
    "RET508",
    "RUF012",
    "UP007"
]

lint.typing-modules = ["colour.hints"]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"

[tool.ruff.lint.per-file-ignores]
"*_test.py" = ["ARG001", "E501", "S101", "PLR2004"]
"test_*.py" = ["ARG001", "E501", "S101", "PLR2004"]
"**/migrations/**" = ["ARG001"]
